refactor: colocate quest-dashboard tests
